+++ dependencies/tools/rstudio-tools.sh
@@ -33,20 +33,6 @@ fi
 export RSTUDIO_TOOLS_ROOT
 
 
-# version of node.js installed with the product
-#
-# In the dependencies folder this will have a `-patched` suffix on the folder name to indicate
-# that some files have been removed (via the patch process mentioned below).
-#
-# When changing node version you must upload the corresponding archives to aws s3; use
-# rstudio/dependencies/tools/upload-node.sh
-#
-# In addition to updating the version here, search the entire repo for other instances of
-# RSTUDIO_INSTALLED_NODE_VERSION and update to match. Also double-check that the steps in
-# patch-node and patch-node.cmd work with the newer node (e.g. in case the layout has changed).
-export RSTUDIO_INSTALLED_NODE_VERSION="20.15.1"
-
-
 # version of go used for building
 export WORKBENCH_GO_VERSION="1.22.3"