+++ src/cpp/session/SessionOptions.cpp
@@ -663,20 +663,6 @@ void Options::resolveCopilotPath(const FilePath& resourcePath,
    }
 }
 
-void Options::resolveNodePath(const FilePath& resourcePath,
-                              std::string* pPath)
-{
-   if (*pPath == kDefaultNodePath && programMode() == kSessionProgramModeDesktop)
-   {
-      FilePath path = macBinaryPath(resourcePath, "node");
-      *pPath = path.getAbsolutePath();
-   }
-   else
-   {
-      resolvePath(resourcePath, pPath);
-   }
-}
-
 void Options::resolveRsclangPath(const FilePath& resourcePath,
                                  std::string* pPath)
 {
@@ -719,18 +705,6 @@ void Options::resolveQuartoPath(const FilePath& resourcePath,
 void Options::resolveCopilotPath(const FilePath& resourcePath, std::string* pPath)
 {
    resolvePath(resourcePath, pPath);
-}
-
-void Options::resolveNodePath(const FilePath& resourcePath,
-                              std::string* pPath)
-{
-#if defined(__linux__) && !defined(RSTUDIO_PACKAGE_BUILD)
-   // node version should match RSTUDIO_INSTALLED_NODE_VERSION
-   FilePath dependenciesPath = resourcePath.completePath("../../dependencies/common/node/20.15.1-patched");
-   resolvePath(dependenciesPath, pPath);
-#else
-   resolvePath(resourcePath, pPath);
-#endif
 }
 
 void Options::resolveRsclangPath(const FilePath& resourcePath,