+++ NEWS.md
@@ -50,6 +50,8 @@
 ### Dependencies
 - Update Electron to version 34.3.1. (#15450)
 - Update Quarto to version 1.6.42. (#15460)
+- Copilot Language Server 1.282.0. (rstudio-pro#7450)
 
 ### Deprecated / Removed
 - No longer building RStudio Desktop or Desktop Pro for OpenSUSE 15, Ubuntu Focal, or RedHat 8. (rstudio-pro/#7445)
+- No longer bundling node.js with RStudio Desktop, Desktop Pro, or RStudio Server. (rstudio-pro#7450)